Split Bug and Comment class out to bug.py from bugdir.py
Comment should probably have it's own file too...

I also tried to clean up the interface for setting status and
severity.  Both attributes involve selecting strings from predefined
lists.  The lists of valid strings (and descriptions of each string)
are now defined in bug.py.  The bug.py lists are then used to generate
appropriate help strings in becommands/status.py and severity.py.
This should make it easier to keep the help strings in synch with the
validation information.

The original status strings weren't documented, and I didn't know what
they all ment, so I elimanted some of them.  'in-progress' and
'disabled' are no longer with us.  Of course, it would be simple to
add them back in if people don't agree with me on that.  Due to the
loss of 'disabled' I had to change the status of two bugs (11e and
597) to 'closed'.  I removed becommands/inprogress.py as well.  It's
functionality was replaced by the more general status.py command,
which mimics the severity.py command.