Explicit rcs.cleanup() in bugdir test.
Don't use del(rcs), because if there was an error, there is still a
reference to rcs in the traceback, so it is never cleaned up.  This
can leave the external archive cluttering up your Arch install if
you're using the Arch backend.  See the __del__ documentation

  http://python.active-venture.com/ref/customization.html#l2h-175

for details.

Also fixed some out-of-date method names in libbe.diff