Added bugid_args option to cmdutil.default_complete.
Now most of the bug-id arguments support Bash completion.  Since there
will hopefully be lots of bugs in the database, I decided to filter
the list of available bugs.  Currently, we just auto-complete active
bugs for most commands, with the exceptions of open (obviously) and
status (which needs to work on all types of bugs).