Added grepm as a template for an output-mbox-to-mutt script.
The goal is to turn grepm into catmutt, and end up with
  $ be --xml list | be-xml-to-mbox | catmutt
to browse current bugs and comments in mutt.

Moritz has generously donated grepm under GPLv2.  Not GPLv>=2 yet, so
if the project decides to go to GLPv3 for example, this file will have
to stay behind.  Not that I see such a change coming, but I thought it
was worth commenting on, so we don't forget.