Add ability to list all current targets.
From his Wed, 24 Jun 2009 23:08:25 +0200 email to the list:

Hello

As I noted some time ago, there is not a way to list all the present targets
in the current repository.

At the time, Chris send me a patch to have this feature, but it still not
merged in the last revision (314).

So this is a patch for this feature to apply against revision 314. It is
basically a port of the old Chris's patch.

bye
Gianluca