Added ability to handle non text/* MIME types.
The main problem was the encoding/decoding that was happening to _all_
input/output.  Now many I/O activities have a `binary' option to
disable any encoding/decoding.  The `binary' flag is set whenever the
comment content-type is not a text/* type.

In order to print valid XML (and make life easy on xml/be-xml-to-mbox),
non text/* types are printed out as base64-encoded MIME messages, so
  be list --xml | be-xml-to-mbox | catmutt
works as you'd expect.

With the standard (non-XML) output from `be show', we just print a
message telling the user that we can't reasonably display the MIME
type and that they should use the XML output if they want to see it.