Updating "be set --help" and "be status --help".
I don't really like the "defaults to None" for the settings that have
funky initialization procedures (most of them :p), but I'm not sure
how to handle that cleanly yet.  Perhaps
  be set --current
I also need to find a method of adding complicated settings like the
nested lists for severities, etc from the "be set" commandline.