be-mbox-to-xml passes attributes on to each part of multipart messages.
Previously "message[<some-attr>]" just returned None if it wasn't set
for that message part, which overwrote anything passed in through
fields.

"from" and "date" added to list of attributes passed along.

For be-xml-to-mbox, "alt-id" was added to Comment._attrs,
and Comment.print_to_mbox was adjusted to handle the case where
we have no information about the parent bug.

With all of this, I can complete the loop
  be-mbox-to-xml example.mbox | be-xml-to-mbox > example2.mbox
without errors :p.

Finally, be-xml-to-mbox has been adjusted to also work on files (it
had previously only handled data via stdin).  We can't add stdin
handling to be-mbox-to-xml though, because the mailbox package needs
an actual file to work on, and I haven't setup a tmpfile workaround
yet...