Added _procmailrc file for interactive email handling.
The beginnings of an interactive email interface to BE.

With a working procmail setup, copying _procmailrc to ~/.procmailrc
should sort through incoming email to that user, passing all messages
with subjects starting with [be-mail] on to the script be-handle-mail
and deleting the rest.

Now I just need to write be-handle-mail ;).