Adjust update_copyright.py to handle _mailfilterrc and _procmailrc
Gour re-released _mailfilterrc under GPLv2+ following Chris'
suggestion:

  Date: Fri, 25 Jun 2010 14:23:46 -0400
  From: Chris Ball
  Subject: Re: [Be-devel] procmail & maildrop
  ...
  As an aside, the Public Domain license on it is mildly concerning,
  but I think I'm willing to take it anyway since it's just a
  standalone file.  There are guidelines against using Public Domain
  declarations for software, because they have different meaning in
  different areas, and some areas are even thought to deny disclaiming
  "moral rights" like this; I've heard that France is in this
  situation.
  ...

Seems reasonable, so I'm doing that for my _procmailrc too.

This also means that _mailfilterrc no longer needs to be in
update_copyright.IGNORED_FILES, so I removed it.  Actually,
since it lackes a '^# Copyright' line, it never had to be
there in the first place.  Lazy me ;).