Run cmd.cleanup_now() after executing bzr commands.
Otherwise
  be remove ...

blocks if it needs to remove multiple files, since
bzrlib.builtins.cmd_remove needs write locks, and the second remove
will try to aquire the lock that the first aquire hadn't released.  If
we force the release, the lock will be available for the second (and
later) removal.

It's not a problem to call cleanup_now() too often, because calling it
clears the cleanup command stack, so I just added explicit cleanups
after every bzr .run() call.