Make libbe.storage.vcs.darcs.Darcs._vcs_listdir() more robust.
The old version returned [] (for Darcs 2.5) on
  darcs show files --no-files --patch 'Initial commit' .be
(called in `be diff` for `test_usage.sh darcs`), because darcs
returned the paths prefixed with './' (e.g. `./.be`, not `.be`).  By
calculating relative paths and using the relative paths to determine
which files belong to the directory, we can handle both prefixed and
plain paths.